+++ FodyWeavers.xml
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
-	<Costura IncludeAssemblies="System.*|Microsoft.*|ControlzEx|MahApps.Metro|MahApps.Metro.IconPacks.Core|MahApps.Metro.IconPacks.Unicons|PeNet*" />
+	<Costura />
 </Weavers>
