+++ packages.config
@@ -2,15 +2,15 @@
 <packages>
   <package id="ControlzEx" version="5.0.1" targetFramework="net462" />
   <package id="Costura.Fody" version="5.7.0" targetFramework="net462" developmentDependency="true" />
-  <package id="Fody" version="6.6.3" targetFramework="net462" developmentDependency="true" />
+  <package id="Fody" version="6.6.4" targetFramework="net462" developmentDependency="true" />
   <package id="MahApps.Metro" version="2.4.9" targetFramework="net462" />
-  <package id="MahApps.Metro.IconPacks" version="4.11.0" targetFramework="net462" />
+  <package id="MahApps.Metro.IconPacks.Unicons" version="4.11.0" targetFramework="net462" />
   <package id="Microsoft.Diagnostics.Tracing.TraceEvent" version="3.0.5" targetFramework="net462" />
-  <package id="Microsoft.NETCore.Platforms" version="6.0.5" targetFramework="net462" />
+  <package id="Microsoft.NETCore.Platforms" version="6.0.6" targetFramework="net462" />
   <package id="Microsoft.Win32.Primitives" version="4.3.0" targetFramework="net462" />
   <package id="Microsoft.Xaml.Behaviors.Wpf" version="1.1.39" targetFramework="net462" />
   <package id="NETStandard.Library" version="2.0.3" targetFramework="net462" />
-  <package id="PeNet" version="2.9.8" targetFramework="net462" />
+  <package id="PeNet" version="2.9.9" targetFramework="net462" />
   <package id="PeNet.Asn1" version="2.0.0" targetFramework="net462" />
   <package id="System.AppContext" version="4.3.0" targetFramework="net462" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net462" />
